requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
av==16.1.0
babel==2.18.0
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.5
csvw==3.7.0
ctranslate2==4.7.1
dataclasses-json==0.6.7
distro==1.9.0
dlinfo==2.0.0
dnspython==2.8.0
ecdsa==0.19.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
faiss-cpu==1.7.4
fastapi==0.109.0
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-crc32c==1.8.0
google-genai==1.63.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.2
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.83
langchain-ollama==0.3.10
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.7.1
language-tags==1.2.0
livekit==1.1.2
livekit-api==1.1.0
livekit-protocol==1.1.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.2
numpy==1.26.4
ollama==0.6.1
onnxruntime==1.23.2
openpyxl==3.1.5
orjson==3.11.7
packaging==25.0
pandas==2.3.3
passlib==1.7.4
phonemizer==3.3.0
pillow==12.1.0
piper-tts==1.4.1
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
PyJWT==2.11.0
pylibsrtp==1.0.0
PyMySQL==1.1.2
pyOpenSSL==25.3.0
pyparsing==3.3.2
pyreadline3==3.5.4
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.3
rdflib==7.5.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
sacremoses==0.1.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
segments==2.3.0
sentence-transformers==2.2.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
SQLAlchemy==2.0.46
starlette==0.35.1
sympy==1.14.0
tenacity==9.1.4
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.13.3
tomli==2.4.0
torch==2.10.0
torchaudio==2.10.0+cpu
torchvision==0.25.0
tqdm==4.67.3
transformers==4.30.2
typer==0.24.1
typer-slim==0.21.2
types-protobuf==6.32.1.20260221
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.27.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
